[backend] Enhance error handling in GET /listing/:listing_id
Modified the route handler for GET /listing/:listing_id to improve
error handling for the case where the listing you want to get the
data for does not exist.

Pulled out the input validation for the path parameter listing_id
into a middleware so it can be reused for PUT and DELETE endpoints.