[backend] Remove outdated server folder
The old server folder was renamed cinder-server due to an issue with
deploying to fly.io which takes the folder name to be the project
name. Removed the server folder in this commit as we now use the
cinder-server folder and this was accidentily commited.

I've moved the changes Peter made within server to the server.ts
file within cinder-server